mm/filemap: support cache page promotion in generic file writes
We know that AEP performs worse for writes when comparing to reads,
whether it be latency or bandwidth.

This patch hard code that knowledge into code, which can be unacceptable
for the community. So it's a quick hack to test out performance of a
simple AEP specific policy.

The more comprehensive solution may be to

- account page read/write hotness
- add weight of HMAT metrics (btw does HMAT distinguish reads/writes?)
- do promotion based on the final weighted hotness score
  (likely different threshold for reads/writes)

which would require a lot more works. So the current patch looks more
like experimental hack than upstream-able solution.

Suggested-by: Fengguang Wu <fengguang.wu@intel.com>
Signed-off-by: Jingqi Liu <jingqi.liu@intel.com>