mm/huge_memory.c: use head to emphasize the purpose of page
During split huge page, it checks the property of the page. Currently we
do the check on page and head without emphasizing the check is on the
compound page. In case the page passed to split_huge_page_to_list is a
tail page, audience would take some time to think about whether the
check is on compound page or tail page itself.

To make it explicit, use head instead of page for those checks. After
this, audience would be more clear about the checks are on compound page
and the page is used to do the split and dump error message if failed.

Signed-off-by: Wei Yang <richardw.yang@linux.intel.com>