maple_tree: remove special handling for mas_start == 0
Currently, on each call mas_mab_cp() we does a check on whether
mas_start == 0. Let's move the check when we find a 0 pivot.

Let's do an approximate calculation. Assume there are 100 times of
mas_mab_cp() call.

Currently, it would to 100 times check.

Since mas_mab_cp() is used to copy a maple node to maple big node, we
usually call 2 times of mas_mab_cp() for one maple node. We can assume
in the total 100 times, 50 times with mas_start == 0 and 50 time not.

Signed-off-by: Wei Yang <richard.weiyang@gmail.com>