maple_tree: leverage wr_mas->r_min to simplify mas_store_b_node()
After tree walk, wr_mas->r_min is used to store the beginning where the
slot points to. This value is usually correct, except for extend
spanning null. By updatint wr_mas->r_min when extend spanning null, we
can unify the handling of the store of first part.

Since the value is already calculated, it is not expected to be worse
than current behavior.

Signed-off-by: Wei Yang <richard.weiyang@gmail.com>
CC: Liam R. Howlett <Liam.Howlett@Oracle.com>
CC: Sidhartha Kumar <sidhartha.kumar@oracle.com>
CC: Lorenzo Stoakes <lorenzo.stoakes@oracle.com>