maple_tree: even a leaf node should update gap
Generally, when a leaf node is changed, we should update the gap.

Current code works well, because mas_wmb_replace() is only called by
mas_spanning_rebalance() and mas_split(). Usually after rebalance and
split, the node is not a leaf node. And the only case for a leaf node is
a new root, in which we don't need to update gap.

Signed-off-by: Wei Yang <richard.weiyang@gmail.com>
CC: Liam R. Howlett <Liam.Howlett@Oracle.com>
CC: Sidhartha Kumar <sidhartha.kumar@oracle.com>
CC: Lorenzo Stoakes <lorenzo.stoakes@oracle.com>