maple_tree: stop destroy walk on hitting 'start'
Current destroy walk is done in a depth first order, which means it only
hit 'start' once and on hitting it the whole subtree has been iterated.

With this knowledge, we can do mt_free_bulk() all in loop and only free
'start' itself in free_leaf.

This also remove the check of (mte_to_node(enode) == node) in each
iteration.

Signed-off-by: Wei Yang <richard.weiyang@gmail.com>
CC: Liam R. Howlett <Liam.Howlett@Oracle.com>