maple_tree: get topiary subtree by comparing with mas's index/last
If an old node's range is in the range being modified, it is the root of
the subtree to be topiary. This range is stored in mas and not changed.

It looks reasonable to compare the range directly with mas instead of
tmp_next, whose range is copied from mas.

Signed-off-by: Wei Yang <richard.weiyang@gmail.com>
CC: Liam R. Howlett <Liam.Howlett@Oracle.com>
CC: Lorenzo Stoakes <lorenzo.stoakes@oracle.com>
CC: Sidhartha Kumar <sidhartha.kumar@oracle.com>